refactor: Move handleClick function to child
With the removal of the Radio button,
it is no longer necessary to have this function in the parent
